Move fixture to conftest.py so it is accessible outside of test_columnar_join
